adds .skip and .only functionality to describeWithDOM
