+++ src/__tests__/Utils-spec.js
@@ -13,9 +13,11 @@ import {
   mapNativeEventNames,
 } from '../Utils';
 import {
-  describeWithDOM,
   mount,
 } from '../';
+import {
+	describeWithDOM,
+} from '../describeWithDOM';
 
 describe('Utils', () => {
