+++ src/index.js
@@ -9,27 +9,7 @@ import { renderToStaticMarkup } from './react-compat';
  * @class Enzyme
  */
 
-let jsdom;
-try {
-  require('jsdom'); // could throw
-  jsdom = require('mocha-jsdom');
-} catch (e) {
-  // jsdom is not supported...
-}
-
 export let sinon = Sinon.sandbox.create();
-
-export function describeWithDOM(a, b) {
-  describe('(uses jsdom)', () => {
-    if (typeof jsdom === 'function') {
-      jsdom();
-      describe(a, b);
-    } else {
-      // if jsdom isn't available, skip every test in this describe context
-      describe.skip(a, b);
-    }
-  });
-}
 
 export function useSetStateHack() {
   let cleanup = false;
